build: remove example-module from gitignore
In order to work around an Angular Bazel bug where genfiles aren't
properly picked up in the  metadata bundler, we need to
check-in the example module into the source-file and use the
example-module as a real source instead of generating it dynamically
through a genfile. This allows us to temporarily fix the examples package
until the bug (angular/angular#30259) is fixed.